feat(IHM): Use module's changelog property (refs #4346)
The use of the <changelog/> node is deprecated, and replaced by a "changelog"
property (on the <module/> node) which contains a URL to the changelog.

The changelog, if it exists, must be specified in the module's "changelog"
property.

E.g. <module name="xxx" changelog="http://example.net/module/xxx/changelog/v1.2.3" />

The web UI opens the changelog URL in a new window when the changelog icon is
clicked.