feat(log): Improve error logging (refs #5080)
HTTP responses' errors are now logged with syslog.

Syslog messages are sent with identity/tag "dynacase-control", facility
LOG_USER and varying priorities (LOG_ERR, LOG_INFO, etc.).

Additionally, if the "Local log" is activated in the Web UI, then
errors are also logged in a `$WIFF_ROOT/log/wiff.log` file. A set of
"View" and "Clear" buttons are available for respectively viewing the
file (as plain text) and clearing it's content.

Also, some system() calls have been rewritten with exec() in order to
capture a more complete output of the commands' errors.