feat(context): Run module's pre-delete phase when deleting a context (refs #4588)
The <pre-delete/> processes of the context's module are now executed
when deleting a context.

This allows a module to perform additional cleanup on the database for
exemple.

The modules are processed by reverse order of theirs dependencies.  So,
the least dependant modules are executed first, and dynacase-core is
treated last.

`xsd` subdir has been updated with support for the `<pre-delete/>` tag.