fix(Module): Use an external entity loader for schema validation (refs #5054)
Setting the catalog file with the XML_CATALOG_FILES env var at runtime
is not supported on Ubuntu 14.10 as the libxml initialization seems to
be done earlier, before our putenv() is executed.

So, to workaround this, we manually parse the catalog.xml file and setup
a custom external entity loader to resolve URNs to local XSD files.

Also:
- Add requirement for PHP >= 5.4 (for libxml_set_external_entity_loader)