fix(Timeouts): Changed handling of connections timeouts (refs #4547)
The timeouts have been reworked and do not try to mimic wget's logic (as
the cURL's timeout do not have the same semantic).

Now, the only configurable timeout is the time to establish a
connection, for fetching a network resource. A new parameter
'connect-timeout' has been introduced to specify in seconds this connect
timeout. Default value is 3 seconds, and a value of 0 keep the default
cURL behaviour.

Also:
- The internal retry logic has been removed, in order to minimize long
  waits from the client side.
- Accessing an invalid repository will now set a mark on it to
  accelerate subsequent access by directly returning an error (instead
  of continuously trying to fetch new contents from this repository).