fix(globalwin): Correct "global" window size problems (refs #5796)(refs #3889)
The globalwin's width and height was constructed from the size of the
context-list element which in some circumstances (when not shown or
properly selected) will return a size of 0.

So, the globalwin's size is now constructed from the size of the
document's body: `Ext.getBody()` instead of
`ext.getCmp("context-list")`.