fix(CLI): Correctly handle the case of contexts being restored (refs #6038)
When listing the contexts, correctly handle the case of contexts being
restored.

By default, contexts being restored are not shown. To show all contexts
(existing contexts with contexts being restored), you'll have to use the
`--all` flag:

    # /var/www/dynacase-control/wiff list context --all --pretty
    Name               Description
    -----------------------------------------------------------------------------------
    c1                 Foo
    c2                 Bar
    * test             Restoration in progress (started on 2016-04-12 14:26:54)

Contexts being restored will be shown prefixed with an asterisk ("*")
and the description will contains the start date and time of the
restoration process.