feat(wstart): Run `wstart -m` on archive creation/restore (refs #6378)
With `dynacase-core` < 3.2.20, running `wstart` would do additional
processing that could produce a context with a content different from
what was originally archived.

So, the `-m` argument was added to `wstart` in `dynacase-core` >= 3.2.20
to only remove the maintenance mode. So, control will now run `wstart
-m` instead of plain `wstart` to remove the maintenance mode without
touching any other files when archiving or restoring a context.