fix(CLI): Cleanup invalid/unsupported chars in parameter's value (refs #6975)
In CLI, control-chars could be entered in the parameter's value and
would produce an XML file that could not be loaded back as it now
contained invalid XML chars.

As a mitigation, the XML invalid chars are now converted to their
corresponding pictogram from the Control Pictures Block or a generic
Unicode replacement symbol.