feat(cli): add support for HTTP Basic auth with A4PPM registry
The repo/registry's credentials are now saved in a XML file named
`.anakeen-cli.credentials` instead of the `repo.xml`.

The order of search for the `.anakeen-cli.credentials` file is as
follow:
1) the file is searched in the current working directory;
2) if not found, the file is searched in the user's home directory;
3) if not found, the file is recursively searched starting from the
current working directory up to the filesystem's root.

If the `.anakeen-cli.credentials` does not exists, it will be created in
the user's home directory.