fix(te): add VERSION and RELEASE files on build (refs #620)
The VERSION and RELEASE file are needed by the `INFO:SERVER` command to
retrieve the transformation server's version from Anakeen Platform's
admin center.

So, we need to programmatically add VERSION and RELEASE files when
producing the packages.

Note that the RELEASE content will be forced to "0" as there is no
explicit "release" semantic anymore.