feat(tools): allow running `make env-start` without a TTY (refs #1119)
`docker-compose exec` requires, by default, a TTY. If the parent process
has no TTY, then we should call `docker-compose exec -T` instead to not
allocate a TTY and prevent `docker-compose` from returning an error "the
input device is not a TTY".

To circumvent this, and allow execution of `make env-start` from a
program without a TTY, we dynamically define the DOCKER_COMPOSE_CMD_EXEC
variable with either `docker-compose exec` or `docker-compose exec -T`
according the presence of a TTY in the calling process.